Refactor commodore.gitrepo into a package
